+++ CHANGELOG.md
@@ -1,8 +1,17 @@
 # Campaign Changelog
 
+## Roadmap
+- Review queue job implementation
+
 ## Unreleased
-- Preflight sendout checks
+
+## 1.0.0-beta1.1 - 2018-03-12
+### Added
+- Added preflight sendout screen
+### Changed
 - Changed "Settings" to "Plugin Settings"
+- Changed "userId" to "senderId" in Sendouts table
+- Moved "actionVerifyEmail" into TrackerController
 
 ## 1.0.0-beta1 - 2018-03-07
 - Initial release of public beta